General change in the way function results are handled. This makes way for better handing of extended function results, and also simplifies function result handling in general.
The function result is moved out of the stack marker for procedure/
function calls. Instead, the allocaton becomes:

1. The function result.
2. The mark.
3. The parameters.
4. local variables.

In stacking order. The function result is now what is left on stack
after the mark is removed. We move the allocation of the function
result to the mst instruction as a parameter, which it does before
creating the mark. The function result can be any size, as aligned
with the stack quanta.

Besides enabling N length function results, this also removes the
need to adjust for the location of the result on 64 bit procesors.

Recompile.

Recompile.

Rerun regression.

1. Removed MARKFV from 64 bit mark definition.

2. Adjusted mark offset values.

3. Fixed mst instruction to allocate function result separately from
mark.

4. Fixed up the ret instructions for the new plan.

1. Removed markfv from 64 bit mark definition.

2. Adjusted mark offset values.

1. Removed the complex calculation for function result location from
selector(). Now is just 0 offset to the mark.

2. Find result size at one place for callnonstandard().

3. Pass in function result size to mst instruction.

1. Add function result length to mst instruction.

2. Refactor ret instructions. Most are just duplicates of each other
now.

Copied changes from pint above.