Further changes for unlimited function returns.
I backed out the previous method. We were heading down the path of
manipulating the stack using primitives from the front end. This is
done elsewhere (dup,mov,swp are examples) however, I believe this is
the wrong technique. Instead we generate fairly high level instructions and let
the machine manage its own affairs.

Thus, we follow the example of sets for the implementation of unlimited
function returns. We generate a stom instruction, that carries both the size of
store and the size of the aligned stack alocation, then it takes care of getting
the destination pointer under the stacked element, moving the stacked result
to the destination, and removing both the stacked structure and the destination
pointer from stack. In:

stom size asize

We pass both the byte size of the result and the stack aligned size. Why?
because both are needed, and I don't think it is the job of the byte machine
to align operands. That is up to the front end.

This will need to be followed by sto instructions for other combinations as
well (local, indirect, etc).

New regression.

1. Removed new instructions (lsa, cpp) we are not going to use.

2. Store refactored to give new stom instruction.

3. assignment() now processes expression results of any length as loads.

4. Assignment() proceses whole array assigns by store(). Note this will need
to be done with records as well.

1. Remove lsa instruction.

2. Add stom instruction.

3. Add retm instruction. retm is the same as other rets, but takes a parameter
indicating the length of the return, which is not used.