Add endian mode compensation.
The existing P6 would use whatever endian mode the host processor
had. It now determines its own endian mode, and then compares it
to the mode of the target. If not the same, the endian mode is flipped.

Regression rerun.

Added endian target mode flag.

1. Added determination of the endian mode of the host procesor. This
is compared to the target, and flipend is set if they are different.

2. Based on flipend, the layout of integers, reals and addresses are
flipped end to end from the host memory mode.

Endian mode compensation is a bit quirky. Normally P6 will adapt to
and produce decks that match it's native endian mode by defintion.
P6 can produce a deck for use on a machine with the opposite endian mode,
and that will run on pint, but it won't run on pmach or cmach unless they
are recompiled with different flags. However an opposite endian mode
generated deck will work on pint, pmach, or cmach on the endian mode
machine it is built for. This allows P6 to be a cross compiler to other
endian machines.