16 bit mode for GPC.
First cut at 16 bit "stepdown" mode for GPC. This is done by a combination of
16 bit MPB definitions and use of fixed length declarations shortint and short
real, which are 16 bit and 32 bit definitions, respectively.

Just contains a commented out 16 bit set. Will rework this into a permanent
solution.

Reran regression.

1. Added endian flag.

2. Reworked the mark declarations.

Added 16 bit define to check for no word size declaration. I.e., it will only
set a default if none of the 16 bit, 32 bit or 64 bit definitions are set.

1. Fixed default as in pcom.pas.

2. Implement special size definitions for shortint, shortreal, and maxint
if the target is GNU Pascal and the size is 16 bit.

3. Redefine code array size definitions for 16 bit mode.

So the "stepdown" [1] definitions for GPC 64 bit should be 32 bit and 16 bit.
As any Pascal compiler should do, integer sizes for machine word type, and so
GPC on 64 bit has a 64 bit integer [2]. We use the "size specific" definitions
of GPC, shortint and shortreal to gain a fast conversion to 16 bit.

Following those same size specific definitions, medint "should" have yielded
a 32 bit integer, and so making a 64 to 32 bit GPC stepdown possible. However,
it didn't work, as verified by testing GPC on 64 bit [3].

So is a 64 to 32 bit stepdown possible? Perhaps, I would need to verify that
a definition like integer = -2147483648..2147483647 works correctly. That is
actually the correct way to do integer subsets in Pascal [4].

[1] my term for a byte machine running with a smaller word size than the host,
i.e., 16 bit words on a 64 bit or 32 bit host.

[2] Pascal is designed to be independent of machine word size, thus integer
floats in length with the target machine. To be fair, C originally also had
this stated property, that int reflects the native word size. In this author's
opinion, the microprocesor implementors made the wrong decision in moving
towards size specific declarations.

[3] Possibly because the 64 bit compile wasn't done by the GPC group itself.

[4] Which is not to say that GPC's size specific way to do this job is
"incorrect". My personal tastes with Pascal don't go with size specific definitions
for Pascal, which is supposed to be a relatively machine independent language.