16 bit byte machine support.
Changed the way 16 bit and addressing is supported in pint.
pint does not need to change the way all integers and reals
are stored, just the way machine integers and reals are
supported. Thus we create a target type for those, and
convert to and from as required. This is relatively quick
in most architectures.

Rerun regression.

1. Created new internal machine types, pmreal and pminteger. These are used
to form actual machine types.

2. Refactored address type to what is actually used on the machine, and
not integer. The equate to integer was done to accomodate negative offsets
and the fact that an address can actually be maxstr+1. The new type
encompasses all of this.

3. All of the integer and real getters and setters are refactored to go
through pminteger and pmreal.

4. With address type no longer being equated to integer, several var parameter
routine calls needed fixing.