16 bit machine support mode.
Further decoupling of the target machine from the code running it.
Created a machine maxint, pmmaxint, and now only the target machine procedures
use that.

Further research on GPC shows that it is not capable of 64 bit to 32 bit
stepdown using the current implemented method of fixed size declarations.
Did the experiment with seeing if subrange like -32768..32767 results in
smaller integers, it does not, thus that method won't work.

This is not a bad outcome, 64 bit or 32 bit stepdown to 16 bit works, that
is useful, and stepdown of 64 bit to 32 bit is not that useful.

Reran regression.

1. Create pmmaxint, coupled to normal maxint or GPC small integer maxint
as required.

2. Moved use of maxint to pmmaxint as required in various parts of code.

Next is retest 16 bit mode, and regularize builds for different modes.