Procedure/function overloads primary support
Reran regression.

1. Add new overload specific error messages. Note we output the old messages
for non-overloaded parameter lists, new messages for overloads.

2. Added functions, isovlfunc() and isovlproc(), to determine if the target
overloaded routine contains a function or a procedure as part of its overload
group. This occurs because procedure overload lists and function overload lists
are disjoint even though grouped together.

3. Added genmst() procedure, moved mst instruction handling out of gen2(). mst
instruction now uses a label to determine function return space. This is a
consequence of the fact that we don't actually know, in the case of overloaded
functions, what is the function result type until the parameter list is
resolved and the correct function selected.

4. Added "squishy" routines to check types like integer, real and char as intt(),
realt(), and chart(). This is the way IP Pascal does things, and it tends to
make type checking easier, since the routines tolerate things like nils and
subranges. I use these only in Pascaline contexts, but we can slowly roll these
into standard routines as needed.

5. call() gets new parameter to tell if it is called in function or procedure
context, so that callnonstandard() can resolve parameter lists according to this.

6. In callnonstandard(), Added nxtprc() procedure that skips forward in the
overload group list if the current parameter matching is not working (in number
or type). If no match is found, sets fcp to nil, and we use fcps to restore
that value for error processing (a bit hokey I know). Note that this handling
matches overloaded and non-overloaded lists equally well, and we determine
if the routine has overloads to steer the correct error message.

7. In callnonstandard(), set label for mst instruction and resolve after the
parameter list is complete and we have the correct overload. Note we reordered
the function result determination to achieve this.

8. In procdeclaration(), added routine compparmovl() (Compare Parameter Overload)
to check if to parameter lists compare to each other as ambiguous overload
lists. As mentioned elsewhere, this is basically a = b (parameter) if a is
assignment compatible with b or b is assignment compatible with a.

9. In procdeclaration(), fixed virtual and overload mixed error.

10. In procdeclaration(), incoming overloads scan across all definitions in the
overload group looking for ambiguity. Note that we don't really do anything
besides flag an error. The lists are all still collected, meaning that the
resolution is arbitrary according to encounter order.

11. In modulep(), we generate a label that equates to zero to generate
the mst instructions for block starts. This is because they are essentially
procedures, with no function result.

1. Moved mst instruction handling to use label for function result size.

This first group of overload changes lacks:

1. checks for convergence with different modes in parameter lists.

2. Checks for each parameter type. Right now it just checks for matching
numbers of parameters in lists.