Prc/func overload
Reran regression.

1. Added new overload error.

2. Refactored compparam() in procdeclaration() into new routines cmppar()
and cmpparlst(). The routine didn't compare procedure/function parameters
properly, and making them test and return boolean instead of directly flag
an error made them useable from several contexts.

3. Refactored callnonstandard() to use new routines.

4. Added new section to callnonstandard() before the procedure/function
parameter check that checks if the call is overload, then finds the correct
overload if that is so. This means that it will proceed into the same
proc/func handling as before with the correct overload, or with the top
overload if no overload is found. Need to suppress errors here.

5. Refactored procdeclaration() to use the new routines.

This fixes overloads for proc/func parameters (and fixes bugs to boot).
remaining is fixup for general parameters. This is difficult due to the ambiguity
between var and value parameters. I checked the IP Pascal handling for the same
and it is unbelievably complex, so don't want to emulate that.