First set of changes for operator overloads.
Reran regression.

1. 'operator' was misclassified as an attribute, took it out of the list.

2. Added array of operator functions to the display. Each display (level) gets
a possible set of overloads, one per operator. Even though operator overloads
are classified across all scoping levels, they are removed with the scope.
Multiple operator overloads are inserted to the group list just as proc/func
overloads.

3. Refactored display inits into new routine inidsp(). The calls to it just
change the non-default items. This allowed the operator array to be initialized
in one place.

4. Added the operator array to disposal.

5. Added errors to: accept only procedure, function and operator as leader;
reject attributes on operator (none actually apply); missing operator after
'operator'.

6. In procdeclaration(), the code is refactored extensively to treat operator
overloads as equivalent to functions. If the operator array entry in the
display is empty, we pick up a new one, otherwise we use that to add to it's
group overload list. The name of the operator function is given as the
operator, like '+' for add/positivition. We will see later if this causes
issues for the debugger. We don't enter these names into symbols, so there is
no collision.

7. The overload processing was missing from function parameter list handling,
so I refactored the handling of parameter lists from procedures and functions
together as a prep for operator overloads, and this fixes the missing function
overload processing as well.

1. Added new routine getlabe() (getlabel extended). This is used in the block
symbol only, and allows the operator characters as well. As mentioned above,
this may require changes for the debugger.

Remaining to do are:

1. Eval overloads over different levels, not just within one level overload
group.

2. Process against the "native overload" of built in functions. This may
be done with a dummy function list or just code, I haven't decided.

3. Place the actual overload code in expression handling.