Operator overloads
Reran regressions.

1. Added cpp (CoPy Parameters) and cpr (CoPy Result) instructions.

2. Added callop2() routine. This is "like" callnonstandard(), but uses the values
stacked from the expression. This routine takes 2 values off the stack, one of
which is passes as lattr, the other is gattr (or right). Finds the overload,
translates the frame and calls the routine. See notes below.

3. Added isopr() function to test for an operator overload existing for the given
operator.

4. Modified simpleexpression() '+' (binary add) processor. If both sides are
numeric, or if in iso7185 mode, the operation is processed as a real and either
integer side is promoted to real. This is a bit tricky, and means that a
combination of numeric and other type are left to the overload routine.
After integer, real, and sets are applied, operator overloads are checked for
where an error would normally occur. If any are found, callop2() is called and
we process the result. Note that this still leaves convert from integer to real
argument to callop2(), which is not yet done.

Added cpp and cpr instructions to backend.

This is the first functional copy of operator overloads, and supports the '+'
only for testing. Others will be added on this general form.

The frame translation is a special problem. Parameters are processed on the stack
with expressions, but there is a difference between the format of a function
call parameter list and an expression, namely that no frame is thrown during
an ordinary expression, and the result area overlaps the input parameters. This
is ok in ordinary expression processing, because the interpreter removes both
operands from the stack and then puts back the result, which a function call
does not do.

We compensate for the format difference by throwing a frame, then "hoisting"
the parameters from the stack into that frame. Then we "drop" the function
result into place after the call. This copy back and forth clearly is ugly,
but allows operator overloads to work with the current system. We can examine
optimization possibilities later.