Additions to operator overloads
Reran regression.

1. Added diagnostic decoder function for structform entries.

2. Fixpar() was moved up so that both callop2() and callnonstandard() can use it.
It was reparameterized to do this.

3. callop2() was refactored to do coercions on input parameters, such as integer
to real, fixed array to container and vice versa. To do this, we detect if a coercion
will occur, then switch from copying all of the parameters into the new frame (see
previous) to copying them one by one and performing the coercion. This also involves
tracking the parameters in both their uncoerced forms as well as coerced form, since
the math we use to bring them to the top of the stack is dependent on which form
they are in.

4. simpleexpression() + operator was changed to load addresses of structured parameters.

This closes out operator overloads for + on most parameter types. The exception would
be N length function results. Also we need to reject var parameters on operator
overloads.

After the + operator is completed and tested we can implement for all other operators
as well.