Operator overloads
This fix adds the ability to accept N-length function results as
parameters to operator overload functions. In regular functions, this is done
via copyback buffers due to the fact that the parameter is onstack but needs
to be addressed for compatibility with the P6 passing convention.

With operator overloads, the actual parameter is under the frame that is
built for compatibility, so no copyback buffer is needed, we simply take
the address where it lives.

Reran regression.

1. Added lsa (Load Stack Address) instruction. It takes the address as an offset
to the stack pointer and pushes it onto the stack, it is used to take the address
of an onstack parameter.

2. In gen2(), added handling to correctly account for stack changes for new
instructions.

3. In callop2(), added handling for case of N-length parameter onstack and typed
as expr. It takes the address of that and thus it references an effective copy.
As all parameters out of the "compatibility" frame, it gets removed by the cpr
or function adjust instruction.

4. In callop2(), moved cpr stack accounting back to the formal system, see note
below.

5. In simpleexpression(), avoided load for expr typed parameters (parameters that
exist only on the stack - this is how N length function results look after the
called function returns). Thus the loading of the address is differred to
callop2().

Added lsa instruction.

There has been a meme going on for a while that we (more correctly I) avoid using
the stack accounting mechanisim (see mes() and mesl()) when the net stack offset
is complex. Thus mesl() was directly called to fix up stack accounting. This is
actually necessary in places like the case statement. This created issues with
callop2(), since it effectly mixes the two systems. Thus we (I) are getting off
this crutch and back to the organized system.

This fix finishes the modes for operator overloads to the + (binary add)
instruction. What follows is adapting this to other operators.