Operator overloads.
Reran regression.

1. Added error number parameter to binopr(), this allows use in different
contexts.

2. In expression(), guarded int to float convertions to only int/real combinations.

3. In expression(), tagged off the various error cases with binopr().

This finishes binary operator overloads. To go are +/-/not uary, and ^ and []
selectors ( . (dot) is not possible, since it would mask all member accesses).

During overload implementation, a paradox was found. = and <> (equal/not equal)
are valid operations for most types, including pointers and object references,
which are closely aligned to pointers (and must be, otherwise you could not have
lists of objects). Using the "nonambiguous" model, that virtually takes these
operators out of use as overloads.

I solution to this is being examined, which makes overload groups only within
a single scope, but allows overloads to mask scopes underlying, just as named
functions mask such scopes. Will do further investigation on this.