Operator overloads: the overlay fix.
Reran regression.

1. Error discovered in comptype. Didn't properly check for both array types.

2. Moved the parameter matching facilities up: parnum(), partype(), and cmptyp().

3. New routine fndopr2() matches binary operands in expressions and returns
matching operator function or nil.

4. Refactored simpleexpression() +/-. Now checks for operator overloads first,
then falls back to system definition. The code for the latter has been mostly
restored to the way it was in P5.

Next is use the discovered operator function in callop2(), and then refactor
the rest of the cases and test the overlay method.