Operator overloads.
Overlay fix.

Reran regression.

1. callop2() refactored to use operator overload function header passed
to it instead of searching for it itself. This reflects the new structure
where the correct overload is found first.

2. selector() refactored to use new callop2() method. This is a traverse
refactor[1].

3. binopr() removed, no longer used.

4. term() got its own copy of fcp (was using the nested copy).

5. term() gets refactored as was simpleexpression(), the test for overloads
first, followed by a return to fairly standard code for the rest.

6. simpleexpression() got reindented, ignore.

7. expression() is refactored along same lines. It was a bit complex, but
makes more sense if you look at the way it was before operator overloads.

Next is refactor of unary operator overloads and fixup selector().

[1] From wikipedia:

"A traverse is a lateral move or route when climbing or descending
(including skiing); going mainly sideways rather than up or down. The general
sense of 'a Traverse' is to cross, or cut across and in general mountaineering,
a road or path traveled traverses the steep gradient of the face."