Operator overloads.
End of overlay fix.

Reran regression.

1. Added fndopr1(), as fndopr1(), but for unaries.

2. callop1() refactored as per callop2().

3. Operator overloads removed from selector(), see below comment.

4. Removed unused vars from callop2().

5. Removed unopr() routine, no longer needed.

6. In factor(), refactored not operator for new overlay scheme.

7. In simpleexpression(), refactored +/- operators.

The ^ and [] operators were removed for the following reasons:

1. Functions cannot return addresses, so any variable reference
using selector would end with an operator function. This then is
hard to explain why such references cannot be chained nor used
as left side of assignments, or used as var parameters.

2. Its generally non-orthogonal, with special rules and one "operator"
that cannot be used, ".".

These operators never made it into the spec, so there is nothing
to do.

This completes the overlay fix. Next is assignment.