Overload operators.
Added assignment operator.

Reran regression.

1. Added operator code for ":=".

2. Added new error for destination in := operator must be "out".

3. Added proc to selector as left side form. This is to handle cases where
a procedure is used as a return value assign. This is flagged as an error
at higher level, and we need selector to process the operands to keep the
parse moving.

4. In callop2(), we handle the procedure case that comes from := operator.

5. Added operator code parameter to parameterlist() so that we can handle
that specially.

6. Procedure and function parameters get no-op as parameterlist() operator.

7. For parameterlist(), we check := operator first list parameter is "out".

8. In procdeclaration(), added := as operator.

9. In procdeclaration(), we add checks for := as a procedure, not a function.

10. Refactor assignment() to check for operator overloads.

1. Added := characters to routine symbols.

Next is add "inherited" function to operator overloads, then copy functionality
to pmach and pint.