Anyedit changes; fix for function overloads of procedures.
Functions that overload procedures weren't being processed correctly,
because it still used the head of the list for reference.

Also added quite a few debugging helpers.

 Changes to be committed:
       modified:   regress_report.txt
       modified:   source/pcom.pas

1. Separated the lex diag from insymbol into prtsym() (print symbol), then
fixed insymbol to call that.

2. Added prtclass() and prtform() to print idclass and structform enumerations.
This replaces previous prtform().

3. Added ovlfunc() function that finds a function in an overload list and
returns that.

4. In chkovlpar(), added check for proc/func/opr status does not match. In this
case, the contexts are distinct, and no parameter matching is required. This is
in the Pascaline standard.

5. In procdeclaration(), when forming the overload, virtual and forward status
of the incoming procedure/function/operator, we consider either functions or
procedures. This is because both can be on the same list. Operators are handled
differently.

6. In statement(), extensively refactored ident processing. Overload lists that
contain both procedures and functions have to be disambiguated, and the
proper overload entry has to be selected for function result assignments.