First part of external routines linkage.
Externals in C (or other language) are done by coupling addresses to a
set of thunks that convert the parameters and return data and call
external routines. Each one must be predefined in the code, thus you
can't just arbitrarily link external routines (this is pretty standard
with interpreters). We take any external module reference and match it
against a set of outside routines by module and symbol. The module is
a standard module, but we hijack any matching names. The module itself
is just used to resolve symbols, and need not be loaded. It is ignored
if it is.

The calls are normal calls, but go to an "extern" map just after the
startup. We call into the map with ordinary cup instructions, but at
the location in the map, a special eext instruction exists, that takes
the location in the map, and form the external call number. Then the
external call handler is executed with the routine number.

The parameters and return are all done via the stack. The thunk takes
parameters off the stack, then calls the target C call. Then it formats
the return value. This is straightforward with value parameters, but
gets more complex for VAR and VIEW parameters.

 Changes to be committed:
       modified:   regress_report.txt
       modified:   source/pint.pas

1 Added new symbols string type, symnam. filnam was too long, the others
too short.

2. Added extvecbase and extvecs variables. These are the base address of
the external vector table, and the length.

3. Add new include, PETIT_AMI. It will probably change to just externals.

4. A new subroutine is added to flabrlc() that converts the names of
the module and the symbol to regular strings and tries to look up the
external. If found, the vector map number replaces the address. This
overrides any module equate.

5. flabrlc() looks up the external routine, and if found, replaces the
address in the resolved cup entry.