2nd part of externals capability.
 Changes to be committed:
       modified:   Makefile
       modified:   gpc/bit32/Makefile

1. Added build of externals C source file.

2. Add EXTERNALS define. Note that at this time, its not really optional.
Like other defines, if turned off it would create missing references.
It needs a "null file" to stub out the calls in case the package does not
exist.

       new file:   source/externals.c

1. Prototype file containing sample calls. This will end up being a series
of includes in the final (petit-ami).

       new file:   source/externals.inc

1. This is the activator file for externals. The calls contained are
lookupexternal(), executeexternal(), and numexternal(). All of the external
calls must be modeled in this file, which right now consists of the petit-ami
set.

       modified:   source/pint.pas

1. Changed the external name from petit-ami to externals, to make it a general
purpose facillity.

2. Moved the include point for externals forward so that the accessor functions
can be used.

3. Moved the ExecuteExternal() call point to the eext instruction handler.
Need to remove diagnostics.