Development of external capability; prevent crash on error.
=== Files changed ===

	deleted:    Makefile
	deleted:    bin/cmach
	deleted:    bin/genobj
	deleted:    bin/pmach
	deleted:    bin/spew
	deleted:    p4/standardpc.cmp
	deleted:    standard_tests/iso7185pat.cmp
	deleted:    standard_tests/iso7185patc.cmp
	deleted:    standard_tests/iso7185pats.cmp

	modified:   gpc/bit64/Makefile

1. Added linemacro capability (correction of error lines).

2. Further support for externals.

	new file:   petit_ami/services.h
	new file:   petit_ami/services.o

1. These are local copies of petit_ami files. See note 1.

	modified:   regress_report.txt
	modified:   source/externals.c

1. This is a mildly useful file for testing. Probably will get
removed. It was used to test the externals before adding real
lib code, so we redirected the calls to a test .C file.

	modified:   source/externals.inc

1. Added pa_list(), which is the most complex call there is. See
note 2.

	modified:   source/pcom.pas

1. Fix crash bug on error.

2. Fixed bazzare extra block. Seemed to have been a leftover from
a conditional block.

	modified:   source/pint.pas

1. Added charptr, pointer to char, which the gpc externals package
needed. Didn't find a way to encapsulate this. It would require
declaration of the externals as a module, would need to use the gpc
modules format, and create problems with coupling to manipulate the
stack, so not worth it at the present time.

=== Bug numbers affected ===

=== Notes ===

1. Basically this prevents the need to use local paths to my petit-ami
directory. This might be permanent, since P6 does not really need to
include all of petit-ami, just the libs. Note that the .h does not
really have a use at the moment, since we use pascal header modules.
(which are way better and cooler in any case).

2. The filrec layout needed to be fixed up so that the C version and the
Pascal version lined up (had the same offset field addresses and
alignment). The list() call and the allenv() call give you a big,
possibly huge, data structure that is allocated outside of pint, and
use different spacing and formatting in any case (for example sets
in P6 are always 32 bytes at present). So externals.inc calls the C
routine, then goes through the resulting data structure and makes
copies internal to the P machine, including getspc() calls.