Added code for allenv() external call.
=== Files changed ===

	modified:   source/externals.inc

1. Added environment pointer variable.

2. Added movstr() routine to move string into VM.

3. Refactor list() routine to use movstr().

4. Added and tested code to thunk allenv().

=== Bug numbers affected ===

=== Notes ===

At this point, we just need the 3 more flavors of exec() to finish
the module. We also need to recycle to C side entries that we
copied.

I considered adding list disposal routines to the services lib, but
honestly, disposing of lists is so simple (and OS independent) it
does not make sense.