pgen development; fixed container problems; reran regression.
 Changes to be committed:
	modified:   doc/the_p6_compiler.docx

1. Filled out intermediate descriptions.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/pcom.pas

1. Changed containerbase() to work on both arrays and arrayc. This was done
because in assign of complex to complex, we convert the fixed to complex in a
mixed operation, but one side, we don't know which, is still an arrays.
This might call into question the name of the routine.

	modified:   source/pint.pas

1. The vin instruction was putting the template lengths into the wrong order.

Notes:

Working to finish all of the outstanding Pascaline intermediate instruction
descriptions.