Turn on all ISO 7185 check flags (that don't compromise operation).
P6 had several ISO 7185 checker flags that were turned off by default. I turned
them on, and assume that program that can't handle that override the option. The
most common of these is the $u flag, which checks undiscriminated variants
(variants without a tag field), because some programs use this feature as a
defacto type caster (including pint). The only ISO 7185 flag required to pass
the PRT that is not set is the $p or "check reuse" flag because it does not
completely recycle returned blocks, and thus can degrade program operation.

Its always going to be true that the user will want to turn off "expensive"
checks like VAR and with block checking, and undefineds checking. This change
just makes that explicit.

 Changes to be committed:
	modified:   basic/basic.pas
	modified:   p4/pcom.pas
	modified:   p4/pint.pas
	modified:   sample_programs/pascals.pas

1. Add $u- flag, allow undiscriminated variants for typecasting.

	modified:   bin/runprt

1. Add noerrmsg flag on runs to supress compiler fail messages. The .err file
is sufficient for testing.

	modified:   doc/the_p6_compiler.docx
	modified:   regress_report.txt

	modified:   source/pcom.pas

1. Turn $i (check VAR blocks) on by default.

2. Turn $u (check undiscriminated variants) on by default.

Note:

Regress does not yet run, p4 has an issue after turning the flags on.