Fix bug with variable block checks enabled but debug checks are not.
 Changes to be committed:
	modified:   regress_report.txt

	modified:   source/pcom.pas

1. The debug checks flag causes tagfields to have their address loaded even if
the location is static. This is because the ensuing ivt instruction needs the
address of the tag. It turns out the nearly identical function of cvb needs this
as well, so we enable ion on either debug checks or variable block checks.

	modified:   source/pint.pas

pint is functionally unmodified, but there are extra debug statements I wanted
to keep.

1. pcs (pc saved) is moved up so we can use it in a routine.

2. dmpins() is added, does a simplified instruction dump.

3. Add simplified istruction dump to faults, disabled by default.

Note:

This issue exists on Pascal-P5 as well.