Fixed parameter size calculation for sets.
 Changes to be committed:
	modified:   source/pcom.pas

1. In "skip" calculation for parameter processing in body, code assumed that
everything below set was pointer size (which it would be after alignment). VAR
passed sets break that rule, so the list calc was out of sync with the parameter
base addresses. It only mattered for the VAR block checking algorithim, and that
was found because it indexed undefined space. This bug also exists in P5.