Reran full regression.
Takes 9 hours for full regression!

Note:

pmach and cmach do not completely run PRT. This is because some tests require
options that are not normally set, which are the heap recycle checks ($m) and
ISO 7185 options ($s). This occurs because pmach and cmach only use the machine
binary deck interface from pint.

I believe the best way to solve that is to allow options to be set from the
command line, using the Pascaline command line parsing interface. Thus I am
kicking this into a future fix.