Remove file entry sharing in pint.
 Changes to be committed:
	modified:   source/pint.pas

1. Add new state, fnone, to file states.

2. Initialize all file entries to fnone.

3. Allocate file entries only from fnone entries.

Notes:

File sharing appears to be dead. There does not appear to be a
case for having two files bear the same logical id.

The issue here is that this was acting as a defacto file reuse system. Now its
possible that we can run out of files. It may be required to have a garbage
collection system for files.